Move HabitsTypes.kt to domain/model
For consistency with other domain types, moved the HabitTag and
IsSelected type aliases from domain/ to domain/model/ where all
other domain model types are located.
