+++ feature/habits/domain/src/commonMain/kotlin/space/be1ski/vibits/feature/habits/domain/model/HabitsTypes.kt
@@ -1,4 +1,4 @@
-package space.be1ski.vibits.feature.habits.domain
+package space.be1ski.vibits.feature.habits.domain.model
 
 typealias HabitTag = String
 typealias IsSelected = Boolean
